All college fetch logic ready to go check description
For home i left it as is since not much else testing is needed but in messages the old fetch function is still in place. When we are ready to switch over to the final product, uncomment the new fetch function below that i put and delete the old one currently in use.